monogram and rank indicator as widget
