+++ lib/features/ranking/widgets/ranking_medal.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 
 import 'package:trainlog_app/app/theme/app_colors.dart';
+import 'package:trainlog_app/app/theme/app_theme.dart';
 
 /// The top-three award icon: a trophy for #1 and a medal for #2/#3, coloured
 /// gold / silver / bronze. Shared by the list rows and the user-position badge,
@@ -29,6 +30,46 @@ class RankingMedal extends StatelessWidget {
       rank == 1 ? Icons.emoji_events : Icons.military_tech,
       size: size,
       color: colorOf(rank),
+    );
+  }
+}
+
+class RankIndicator extends StatelessWidget {
+  final int rank;
+  final bool showRank;
+
+  const RankIndicator({super.key, required this.rank, required this.showRank});
+
+  @override
+  Widget build(BuildContext context) {
+    final cs = Theme.of(context).colorScheme;
+
+    if (!showRank) return const SizedBox(width: 34);
+
+    if (RankingMedal.isMedal(rank)) {
+      return Container(
+        width: 34,
+        height: 34,
+        alignment: Alignment.center,
+        decoration: BoxDecoration(
+          color: RankingMedal.colorOf(rank).withValues(alpha: 0.18),
+          shape: BoxShape.circle,
+        ),
+        child: RankingMedal(rank: rank, size: 18),
+      );
+    }
+
+    return SizedBox(
+      width: 34,
+      child: Text(
+        '$rank',
+        textAlign: TextAlign.center,
+        style: AppTheme.monoFont.copyWith(
+          fontSize: 15,
+          fontWeight: FontWeight.w700,
+          color: cs.onSurfaceVariant,
+        ),
+      ),
     );
   }
 }
